fix a sign compare warning.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@10062 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/inet_pton.c
@@ -145,7 +145,7 @@ inet_pton4(src, dst, pton)
 		for (;;) {
 			if (isdigit(c)) {
 				digit = c - '0';
-				if (digit >= base)
+				if ((int)digit >= base)
 					break;
 				val = (val * base) + digit;
 				c = *++src;
